requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bitarray==3.6.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
ckzg==2.1.1
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
cytoolz==1.0.1
distro==1.9.0
docutils==0.22
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fonttools==4.59.1
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
httpx-aiohttp==0.1.8
hyperliquid-python-sdk==0.17.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
msgpack==1.1.1
multidict==6.6.4
narwhals==2.0.1
numpy==2.3.2
od==2.0.2
packaging==25.0
pandas==2.3.1
parsimonious==0.10.0
pillow==11.3.0
plotly==6.2.0
propcache==0.3.2
protobuf==6.31.1
pyarrow==21.0.0
pycares==4.10.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rlp==4.1.0
rpds-py==0.27.0
setuptools==80.9.0
sigtools==4.0.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.48.0
tenacity==9.1.2
toml==0.10.2
toolz==1.0.0
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websocket-client==1.8.0
yarl==1.20.1
